refactor(desktop): split LinkHoverTooltip hook/component, tighten modifier listener
- Move useLinkHoverState into its own hooks/ folder (was exported alongside
  the component in violation of the one-component-per-file rule).
- Effect now re-subscribes on hover start/end only (deps: hovering boolean),
  not on every modifier change.
- Filter to Meta/Control/Shift/Alt key events so typing a letter while
  hovering doesn't churn state.
- Skip setState when modifier/shift values didn't actually change, avoiding
  identity-change re-renders on repeat keydowns.
- Extract tooltip offset constant.
