+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/TerminalPane/components/LinkHoverTooltip/LinkHoverTooltip.tsx
@@ -1,17 +1,12 @@
 import type { ExternalApp } from "@superset/local-db";
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { createPortal } from "react-dom";
 import { getAppOption } from "renderer/components/OpenInExternalDropdown/constants";
 import type { LinkHoverInfo } from "renderer/lib/terminal/terminal-runtime-registry";
 import { electronTrpcClient } from "renderer/lib/trpc-client";
+import type { HoveredLink } from "../../hooks/useLinkHoverState";
 
-interface HoveredLink {
-	clientX: number;
-	clientY: number;
-	info: LinkHoverInfo;
-	modifier: boolean;
-	shift: boolean;
-}
+const TOOLTIP_OFFSET_PX = 14;
 
 interface LinkHoverTooltipProps {
 	hoveredLink: HoveredLink | null;
@@ -64,52 +59,12 @@ export function LinkHoverTooltip({ hoveredLink }: LinkHoverTooltipProps) {
 		<div
 			className="pointer-events-none fixed z-50 w-fit rounded-md bg-foreground px-3 py-1.5 text-xs text-background"
 			style={{
-				left: hoveredLink.clientX + 14,
-				top: hoveredLink.clientY + 14,
+				left: hoveredLink.clientX + TOOLTIP_OFFSET_PX,
+				top: hoveredLink.clientY + TOOLTIP_OFFSET_PX,
 			}}
 		>
 			{label}
 		</div>,
 		document.body,
 	);
-}
-
-export function useLinkHoverState() {
-	const [hoveredLink, setHoveredLink] = useState<HoveredLink | null>(null);
-
-	useEffect(() => {
-		if (!hoveredLink) return;
-		const update = (event: KeyboardEvent) => {
-			setHoveredLink((prev) => {
-				if (!prev) return null;
-				return {
-					...prev,
-					modifier: event.metaKey || event.ctrlKey,
-					shift: event.shiftKey,
-				};
-			});
-		};
-		window.addEventListener("keydown", update);
-		window.addEventListener("keyup", update);
-		return () => {
-			window.removeEventListener("keydown", update);
-			window.removeEventListener("keyup", update);
-		};
-	}, [hoveredLink]);
-
-	const onHover = useCallback((event: MouseEvent, info: LinkHoverInfo) => {
-		setHoveredLink({
-			clientX: event.clientX,
-			clientY: event.clientY,
-			info,
-			modifier: event.metaKey || event.ctrlKey,
-			shift: event.shiftKey,
-		});
-	}, []);
-
-	const onLeave = useCallback(() => {
-		setHoveredLink(null);
-	}, []);
-
-	return { hoveredLink, onHover, onLeave };
 }
