+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/TerminalPane/components/LinkHoverTooltip/index.ts
@@ -1,1 +1,1 @@
-export { LinkHoverTooltip, useLinkHoverState } from "./LinkHoverTooltip";
+export { LinkHoverTooltip } from "./LinkHoverTooltip";
