+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/TerminalPane/hooks/useLinkHoverState/index.ts
@@ -1,0 +1,1 @@
+export { type HoveredLink, useLinkHoverState } from "./useLinkHoverState";
